2019/04/30 01/39

+++ src/main/java/bohdan/stasiv/buyitem/dto/response/ProductsResponse.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -12,7 +14,7 @@ public class ProductsResponse {
 
     private Long id;
     private String name;
-    private Long price;
+    private BigDecimal price;
     private String categoryName;
 //    private String brandName;
 //    private Double volume;
